+++ test/test_heartbeat_monitoring.py
@@ -20,7 +20,7 @@ import threading
 sys.path[0:0] = [""]
 
 from pymongo.errors import ConnectionFailure
-from pymongo.ismaster import IsMaster
+from pymongo.hello import Hello
 from pymongo.monitor import Monitor
 from test import unittest, client_knobs, IntegrationTest
 from test.utils import (HeartbeatEventListener, MockPool, single_client,
@@ -38,7 +38,7 @@ class TestHeartbeatMonitoring(IntegrationTest):
                 def _check_with_socket(self, *args, **kwargs):
                     if isinstance(responses[1], Exception):
                         raise responses[1]
-                    return IsMaster(responses[1]), 99
+                    return Hello(responses[1]), 99
 
             m = single_client(
                 h=uri,
@@ -62,7 +62,7 @@ class TestHeartbeatMonitoring(IntegrationTest):
                 self.assertEqual(actual.connection_id,
                                  responses[0])
                 if expected != 'ServerHeartbeatStartedEvent':
-                    if isinstance(actual.reply, IsMaster):
+                    if isinstance(actual.reply, Hello):
                         self.assertEqual(actual.duration, 99)
                         self.assertEqual(actual.reply._doc, responses[1])
                     else: